requirements.txt
fastapi==0.100.0
pydantic==1.10.11
python-dotenv==1.0.0
uvicorn==0.23.1
langchain==0.0.240
beautifulsoup4==4.12.2
tiktoken==0.4.0
httpx==0.24.1
openai
chromadb
python-environ
psycopg2
streamlit
pytest